refactor(DataGrid): Adjust some typings
