+++ packages/react-component-library/src/components/DataGrid/DataGrid.d.ts
@@ -1,8 +1,11 @@
 import '@tanstack/react-table'
 
+import { TABLE_COLUMN_ALIGNMENT } from './constants'
+import { ValueOf } from '../../helpers'
+
 declare module '@tanstack/react-table' {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   interface ColumnMeta<TData extends RowData, TValue> {
-    align: 'left' | 'right' | 'center'
+    align: ValueOf<typeof TABLE_COLUMN_ALIGNMENT>
   }
 }
