--- packages/react-component-library/src/components/Pagination/types.ts
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export type OnChangeEventType = React.KeyboardEvent | React.MouseEvent
